Separate tests of tutorial dataset and earth_relief dataset into two files
